Remove the KestrelConnection cast - Moved feature implementations into KestrelConnection out of TransportConnection
