+++ src/Servers/Kestrel/Transport.Abstractions/src/Internal/TransportConnection.FeatureCollection.cs
@@ -21,16 +21,11 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Abstractions.Internal
                                                IConnectionItemsFeature,
                                                IMemoryPoolFeature,
                                                IConnectionLifetimeFeature,
-                                               IConnectionHeartbeatFeature,
-                                               IConnectionLifetimeNotificationFeature,
-                                               IConnectionCompleteFeature
+                                               IConnectionLifetimeNotificationFeature
     {
         // NOTE: When feature interfaces are added to or removed from this TransportConnection class implementation,
         // then the list of `features` in the generated code project MUST also be updated.
         // See also: tools/CodeGenerator/TransportConnectionFeatureCollection.cs
-
-        private Stack<KeyValuePair<Func<object, Task>, object>> _onCompleted;
-        private bool _completed;
 
         string IHttpConnectionFeature.ConnectionId
         {
@@ -91,87 +86,5 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Abstractions.Internal
         void IConnectionLifetimeFeature.Abort() => Abort(new ConnectionAbortedException("The connection was aborted by the application via IConnectionLifetimeFeature.Abort()."));
 
         void IConnectionLifetimeNotificationFeature.RequestClose() => RequestClose();
-
-        void IConnectionHeartbeatFeature.OnHeartbeat(System.Action<object> action, object state)
-        {
-            OnHeartbeat(action, state);
-        }
-
-        void IConnectionCompleteFeature.OnCompleted(Func<object, Task> callback, object state)
-        {
-            if (_completed)
-            {
-                throw new InvalidOperationException("The connection is already complete.");
-            }
-
-            if (_onCompleted == null)
-            {
-                _onCompleted = new Stack<KeyValuePair<Func<object, Task>, object>>();
-            }
-            _onCompleted.Push(new KeyValuePair<Func<object, Task>, object>(callback, state));
-        }
-
-        public Task CompleteAsync()
-        {
-            if (_completed)
-            {
-                throw new InvalidOperationException("The connection is already complete.");
-            }
-
-            _completed = true;
-            var onCompleted = _onCompleted;
-
-            if (onCompleted == null || onCompleted.Count == 0)
-            {
-                return Task.CompletedTask;
-            }
-
-            return CompleteAsyncMayAwait(onCompleted);
-        }
-
-        private Task CompleteAsyncMayAwait(Stack<KeyValuePair<Func<object, Task>, object>> onCompleted)
-        {
-            while (onCompleted.TryPop(out var entry))
-            {
-                try
-                {
-                    var task = entry.Key.Invoke(entry.Value);
-                    if (!ReferenceEquals(task, Task.CompletedTask))
-                    {
-                        return CompleteAsyncAwaited(task, onCompleted);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Logger?.LogError(ex, "An error occured running an IConnectionCompleteFeature.OnCompleted callback.");
-                }
-            }
-
-            return Task.CompletedTask;
-        }
-
-        private async Task CompleteAsyncAwaited(Task currentTask, Stack<KeyValuePair<Func<object, Task>, object>> onCompleted)
-        {
-            try
-            {
-                await currentTask;
-            }
-            catch (Exception ex)
-            {
-                Logger?.LogError(ex, "An error occured running an IConnectionCompleteFeature.OnCompleted callback.");
-            }
-
-            while (onCompleted.TryPop(out var entry))
-            {
-                try
-                {
-                    await entry.Key.Invoke(entry.Value);
-                }
-                catch (Exception ex)
-                {
-                    Logger?.LogError(ex, "An error occured running an IConnectionCompleteFeature.OnCompleted callback.");
-                }
-            }
-        }
     }
 }
